+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Fixed `SystemStackError` with `Torch::Tensor#<=>`.
+
 ## 0.19.1 (2025-02-10)
 
 - Fixed error with Rice 4.5